Specificity Functions

siRNATools.Specificity.ReferenceSequence

siRNATools.Specificity.decode_refseq

siRNATools.Specificity.decode_refseq_partial

siRNATools.Specificity.download_RefSeq

siRNATools.Specificity.encode_refseq

siRNATools.Specificity.get_refseq_pos

siRNATools.Specificity.process_RefSeq

siRNATools.Specificity.save_RefSeq

Reference Sequence

siRNATools.Specificity.ReferenceSequence — Type

Structure for reference sequences. Compresses RNA data into 2 bits of information from the 8 of a normal character string. Can only use bases A, C, G, U. A => 00 C => 01 G => 10 U => 11

siRNATools.Specificity.get_refseq_pos — Function

get_refseq_pos(::ReferenceSequence, ::Int)

Function to retrieve bit value of base at specificied position in a reference sequence.

siRNATools.Specificity.encode_refseq — Function

encode_refseq(::String) :: Reference Sequence

Function that takes a string containing only A, C, G, and U and encodes it into a ReferenceSequence type

siRNATools.Specificity.decode_refseq — Function

decode_refseq(::ReferenceSequence) :: String

Function takes a ReferenceSequence type and returns a String containing the RNA bases.

siRNATools.Specificity.decode_refseq_partial — Function

decode_refseq_partial(::ReferenceSequence, ::UnitRange) :: String

Function which takes a ReferenceSequence type and outputs specificied range of bases as a String.

siRNATools.Specificity.download_RefSeq — Function

download_RefSeq(::UnitRange{Int64}=1:8, ::String=PATH)

Downloads mRNA reference sequence from ftp://ftp.ncbi.nlm.nih.gov/refseq/H_sapiens/mRNA_Prot/ to the PATH folder. Defaults to downloading 8 files.

siRNATools.Specificity.process_RefSeq — Function

process_RefSeq(::UnitRange{Int64}=1:8, ::String=PATH)

Processes raw gzipped fasta files into DataFrame and saves it as a CSV in the PATH folder

siRNATools.Specificity.save_RefSeq — Function

save_RefSeq(::String=PATH)

Saves relevant data structures from the processes mRNA reference sequence for use in searches.

TranscriptGene => dictionary of Transcripts to Genes

GeneTranscripts => dictionary of Genes to Transcripts

allT => dictionary of Transcript name to base sequence as String

allRefSeq => dictionary of Transcript name to base sequence as ReferenceSequence

Genome Search