Specificity Functions

siRNATools.Specificity.ReferenceSequence

siRNATools.Specificity.Calculate_Specificity

siRNATools.Specificity.calculate_Peq

siRNATools.Specificity.compress_genome_matches

siRNATools.Specificity.decode_refseq

siRNATools.Specificity.decode_refseq_partial

siRNATools.Specificity.download_RefSeq

siRNATools.Specificity.encode_refseq

siRNATools.Specificity.final_calc

siRNATools.Specificity.find_genome_matches

siRNATools.Specificity.find_match_sequences

siRNATools.Specificity.get_refseq_pos

siRNATools.Specificity.mismatch_positions

siRNATools.Specificity.motif_to_transcript_match

siRNATools.Specificity.process_RefSeq

siRNATools.Specificity.reverse_complement

siRNATools.Specificity.save_RefSeq

Reference Sequence

siRNATools.Specificity.ReferenceSequence — Type

Structure for reference sequences. Compresses RNA data into 2 bits of information from the 8 of a normal character string. Can only use bases A, C, G, U.

A => 00

C => 01

G => 10

U => 11

siRNATools.Specificity.get_refseq_pos — Function

get_refseq_pos(::ReferenceSequence, ::Int)

Function to retrieve bit value of base at specificied position in a reference sequence.

siRNATools.Specificity.encode_refseq — Function

encode_refseq(::String) :: ReferenceSequence

Function that takes a string containing only A, C, G, and U and encodes it into a ReferenceSequence type

siRNATools.Specificity.decode_refseq — Function

decode_refseq(::ReferenceSequence) :: String

Function takes a ReferenceSequence type and returns a String containing the RNA bases.

siRNATools.Specificity.decode_refseq_partial — Function

decode_refseq_partial(::ReferenceSequence, ::UnitRange) :: String

Function which takes a ReferenceSequence type and outputs specificied range of bases as a String.

siRNATools.Specificity.download_RefSeq — Function

download_RefSeq(::UnitRange{Int64}=1:8, ::String=PATH)

Downloads mRNA reference sequence from ftp://ftp.ncbi.nlm.nih.gov/refseq/H_sapiens/mRNA_Prot/ to the PATH folder. Defaults to downloading 8 files.

siRNATools.Specificity.process_RefSeq — Function

process_RefSeq(::UnitRange{Int64}=1:8, ::String=PATH)

Processes raw gzipped fasta files into DataFrame and saves it as a CSV in the PATH folder

siRNATools.Specificity.save_RefSeq — Function

save_RefSeq(::String=PATH)

Saves relevant data structures from the processes mRNA reference sequence for use in searches.

TranscriptGene => dictionary of Transcripts to Genes

GeneTranscripts => dictionary of Genes to Transcripts

allT => dictionary of Transcript name to base sequence as String

allRefSeq => dictionary of Transcript name to base sequence as ReferenceSequence

Genome Search

siRNATools.Specificity.Calculate_Specificity — Function

Calculate_Specificity(patterns, excluded_gene="", rg=2:18, verbose=true) :: DataFrame

Function takes as input the pattern to be searched against the genome, the excluded gene to be ignored, the range of pattern being used, and a boolean which controls whether progress bars will be shown. Output is a DataFrame with a column for the pattern, the number of genes with minimum mismatch distance of 0-4 and the specificity score.

siRNATools.Specificity.reverse_complement — Function

reverse_complement(::String) :: String

Takes the reverse complement of an RNA string

siRNATools.Specificity.calculate_Peq — Function

calculate_Peq(::String) :: Array{UInt64, 1}

Takes as input an RNA strand and translates it into an array of Unsigned Integers each one representing one letter and each bit referring to a position. Used as a pre-processing step for Myers searching algorithm. Maximum length of searched pattern is 64.

Example

AACGCU becomes [110000, 001010, 000100, 000001]

siRNATools.Specificity.motif_to_transcript_match — Function

motif_to_transcript_match(Peq::Array{Unit64, 1}, m::Int64, refseq::ReferenceSequence, min_K::Int) :: Array{Tuple{UInt64, UInt64}}

Takes as input Peq calculated in calculate_Peq, m = the length of the pattern to search for, the ReferenceSequence to search in, and min_k = the maximum distance to output. Implements the string matching algorithm described in A Fast Bit-Vector Algorithm for Approximate String Matching Based on Dynamic Programming by Gene Myers. Output is a tuple of all Levenshtein distances less than min_k and their positions in refseq.

siRNATools.Specificity.find_match_sequences — Function

find_match_sequences(motif::String, sequence::String, mismatches::Int) :: Array{String, 1}

Function takes as input a motif, sequence, and number of mismatches to search for. Output is an array of all substrings of sequence which have a Hamming distance of exactly mismatches to motif.

siRNATools.Specificity.mismatch_positions — Function

mismatch_positions(::String, ::String) :: Array{Int, 1}

Function returns an array of all positions in which strings used as input differ. Strings must be the same length.

siRNATools.Specificity.find_genome_matches — Function

find_genome_matches(pattern::String, excluded_gene::String = "", verbose::Bool = true, minimum_matches = 5) :: Array{Tuple{String, Int64}}

Function takes as input a pattern to search the genome for, excludedgene to exclude the gene of interest from the search, verbose set to true displays a progress bar showing progress of search, and minimum\matches which is the amount of mismatches searched for + 1. Output is an array of tuples of transcript names and the lowest Hamming distance found to the pattern within that transcript.

siRNATools.Specificity.compress_genome_matches — Function

compress_genome_matches(::Array{Tuple{String, Int64}}) :: Dict{String, Array{Int64, 1}}

Function takes as input the output of find_genome_matches and collapses it into a dictionary of Gene names and an array of all lowest Hamming mismatch of the transcripts of that gene.

siRNATools.Specificity.final_calc — Function

final_calc(pattern::String, raw_data::Array{Tuple{String, Int64}}, compressed_data::Dict{String, Array{Int64, 1}})

Function takes as input the pattern being searched for, the raw_data from find_genome_matches, and the compressed data from compress_genome_matches. Ouput is a Tuple containing the mismatch_counts dictionary which adds up the number of genes with a minimum Hamming distance of 0-4, and the specificity score.